# server/common/swagger/Api.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: expressBoilerplate
  description: Express boilerplate
basePath: /api/v1
tags:
  - name: Categories
    description: Categories list
  - name: Specification
    description: The swagger API specification

consumes:
  - application/json
produces:
  - application/json

definitions:
  CategoryBody:
    type: object
    title: category
    required:
      - name
    properties:
      name:
        type: string
        description: The category name

  MovieBody:
    type: object
    title: movie
    required:
      - name
    properties:
      name:
        type: string
        description: The movie name

paths:
  /categories:
    get:
      tags:
        - Categories
      description: Fetch all categories
      responses:
        200:
          description: Returns all categories
    post:
      tags:
        - Categories
      description: Create a new category
      parameters:
        - name: category
          in: body
          description: number of items to skip
          required: true
          schema:
            $ref: "#/definitions/CategoryBody"
      responses:
        200:
          description: Returns all categories

  /categories/{id_category}:
    get:
      tags:
        - Category
      description: Get category information
      parameters:
        - name: id_category
          in: path
          description: Id category
          required: true
          type: integer
          format: int64

      responses:
        200:
          description: Return category info
        404:
          description: Category not found
    put:
      tags:
        - Category
      description: Update category
      parameters:
        - name: id_category
          in: path
          description: Id category
          required: true
          type: integer
          format: int64
        - name: category
          in: body
          description: number of items to skip
          required: true
          schema:
            $ref: "#/definitions/CategoryBody"
      responses:
        200:
          description: Return all codes
        404:
          description: Item not found

    delete:
      tags:
        - Category
      description: Delete category
      parameters:
        - name: id_category
          in: path
          description: Id category
          required: true
          type: integer
          format: int64

      responses:
        200:
          description: Return category info
        404:
          description: Category not found


  /movies:
    get:
      tags:
        - movies
      description: Fetch all movies
      responses:
        200:
          description: Returns all movies
    post:
      tags:
        - movies
      description: Create a new movie
      parameters:
        - name: movie
          in: body
          description: number of items to skip
          required: true
          schema:
            $ref: "#/definitions/MovieBody"
      responses:
        200:
          description: Returns all movies

  /movies/{id_movie}:
    get:
      tags:
        - movie
      description: Get movie information
      parameters:
        - name: id_movie
          in: path
          description: Id movie
          required: true
          type: integer
          format: int64

      responses:
        200:
          description: Return movie info
        404:
          description: movie not found

    put:
      tags:
        - Movie
      description: Update movie
      parameters:
        - name: id_movie
          in: path
          description: Id movie
          required: true
          type: integer
          format: int64
        - name: movie
          in: body
          description: number of items to skip
          required: true
          schema:
            $ref: "#/definitions/MovieBody"
      responses:
        200:
          description: Return all codes
        404:
          description: Movie not found

    delete:
      tags:
        - Movie
      description: Delete movie
      parameters:
        - name: id_movie
          in: path
          description: Id movie
          required: true
          type: integer
          format: int64

      responses:
        200:
          description: Return movie info
        404:
          description: Movie not found
  /spec:
    get:
      tags:
        - Specification
      responses:
        200:
          description: Return the API specification